requirements.txt
Flask==2.0.1
requests==2.23.0
websocket-client==0.53.0